added more types to the types file. renamed index.js to messages.js in dataController folder
