+++ src/components/messages/Block.jsx
@@ -1,19 +1,16 @@
 import { useDispatch } from "react-redux";
 import { types } from "../../redux/types/types";
-import { API_URL } from "../../config/general";
-import { callAPI } from "../../dataController";
-import axios from "axios";
+import { callAPI } from "../../dataController/messages";
 
 const Block = (props) => {
 	const dispatch = useDispatch();
-	console.log(props);
 
 	return (
 		<div className="m-4">
 			<button
 				className="btn btn-danger btn-sm shadow"
 				onClick={() => {
-					callAPI("BLOCK_USER", props);
+					callAPI(types.BLOCK_USER, props);
 					dispatch({ type: types.BLOCK_USER, payload: props });
 				}}
 			>
