+++ src/components/messages/Message.jsx
@@ -1,6 +1,6 @@
 import { useState } from "react";
 import { useDispatch } from "react-redux";
-import { callAPI } from "../../dataController";
+import { callAPI } from "../../dataController/messages";
 import { types } from "../../redux/types/types";
 
 const Message = (props) => {
@@ -29,7 +29,7 @@ const Message = (props) => {
 				<button
 					className="btn btn-sm btn-dark position-absolute top-0 end-0 shadow"
 					onClick={() => {
-						callAPI("DELETE_MESSAGE", { messageId: message.messageId });
+						callAPI(types.DELETE_MESSAGE, { messageId: message });
 						dispatch({
 							type: types.DELETE_MESSAGE,
 							payload: message.messageId,
