+++ src/redux/types/types.js
@@ -1,20 +1,21 @@
 export const types = {
-  DELETE_MESSAGE: "DELETE_MESSAGE",
-  BLOCK_USER: "BLOCK_USER",
-  SET_USER_INPUT: "SET_USER_INPUT",
-  ADD_MESSAGE: "ADD_MESSAGE",
-  ADD_USER: "ADD_USER",
-  ADD_TO_SEEN: "ADD_TO_SEEN",
-  SET_SCREEN: "SET_SCREEN",
-  SET_ONBOARDING_SCREEN: "SET_ONBOARDING_SCREEN",
-  SET_REGISTER_SCREEN: "SET_REGISTER_SCREEN",
-  ON_INPUT_REG: "ON_INPUT_REG",
-  ADD_TO_LIKES: "ADD_TO_LIKES",
-  UPDATE_MATCHES: "UPDATE_MATCHES",
-  SET_REG_ERRORS: "SET_REG_ERRORS",
-  SET_FILTER_OPTIONS: "SET_FILTER_OPTIONS",
-  SET_CURRENT_USER_ID: "SET_CURRENT_USER_ID",
-  SET_LOGGED_IN_STATUS: "SET_LOGGED_IN_STATUS",
-  LOG_OUT: "LOG_OUT",
-  BYPASS_ONBOARDING: "BYPASS_ONBOARDING",
+	DELETE_MESSAGE: "DELETE_MESSAGE",
+	BLOCK_USER: "BLOCK_USER",
+	SET_USER_INPUT: "SET_USER_INPUT",
+	ADD_MESSAGE: "ADD_MESSAGE",
+	ADD_USER: "ADD_USER",
+	ADD_TO_SEEN: "ADD_TO_SEEN",
+	SET_SCREEN: "SET_SCREEN",
+	SET_ONBOARDING_SCREEN: "SET_ONBOARDING_SCREEN",
+	SET_REGISTER_SCREEN: "SET_REGISTER_SCREEN",
+	ON_INPUT_REG: "ON_INPUT_REG",
+	ADD_TO_LIKES: "ADD_TO_LIKES",
+	UPDATE_MATCHES: "UPDATE_MATCHES",
+	SET_REG_ERRORS: "SET_REG_ERRORS",
+	SET_FILTER_OPTIONS: "SET_FILTER_OPTIONS",
+	SET_CURRENT_USER_ID: "SET_CURRENT_USER_ID",
+	SET_LOGGED_IN_STATUS: "SET_LOGGED_IN_STATUS",
+	LOG_OUT: "LOG_OUT",
+	BYPASS_ONBOARDING: "BYPASS_ONBOARDING",
+	GET_USER_MESSAGES: "GET_USER_MESSAGES",
 };
